feat: LiveList 컴포넌트 생성 (+LiveLists -> LiveCollection으로 이름 변경)
