+++ frontend/src/components/LiveCollection/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './LiveCollection';
