--- frontend/src/components/LiveLists/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './LiveLists';
